HUGE Gameplay Update (All implementations in the description)
- "Respawn" button
Now if you press R you will respawn in the last checkpoint.

- Checkpoints
In order to finish a lap you have to pick all checkpoints. If you die or
you use the "Respawn" button you will reappear in the last checkpoint
reached.

In order to finish a lap you have to touch in order all checkpoints. The
next checkpoint is marked with a red laser.

- Death Blocks
You can die now if you fall from the circuit or if you try to cheat
jumping from the high level to the low level.

-Anti-cheat feature
You can't skip checkpoints. You can't jump from the high level to the
low level either.